- Cleaned code - Added the possibility of having CHR-RAM in MMC3 (for Megaman 6) - Added MMC4 (Mapper 10) - Changed analogic signal output from uint8_t to float (for AD1 / AD2 2A03 signals)

+++ NESEmu/NESEmu/Apu/Chip.hpp
@@ -49,8 +49,8 @@ namespace NESEmu { namespace Apu {
         void dmcSampleFetched(uint8_t data);
         
         // Channels output
-        uint8_t getPulsesOutput() const;
-        uint8_t getTriangleNoiseDmcOutput() const;
+        float getPulsesOutput() const;
+        float getTriangleNoiseDmcOutput() const;
         
     private:
